+++ src/Console/Command/CheckCommand.php
@@ -7,6 +7,7 @@ namespace Normalizator\Console\Command;
 use Exception;
 use Iterator;
 use Normalizator\Configuration\Configurator;
+use Normalizator\Finder\File;
 use Normalizator\Finder\Finder;
 use Normalizator\Normalizator;
 use Normalizator\Util\Logger;
@@ -22,7 +23,9 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
+use function array_unique;
 use function count;
+use function file_exists;
 use function implode;
 use function iterator_count;
 use function memory_get_peak_usage;
@@ -72,6 +75,8 @@ class CheckCommand extends Command
 
             The <comment>--middle-eol[=NUM], -m [NUM]</comment> option trims redundant newlines in the middle of the content. The value presents the maximum allowed middle final newlines. Default number of middle newlines is 1.
 
+            The <comment>--not</comment> option(s) will skip given paths for normalizations.
+
             The <comment>--trailing-whitespace, -w</comment> option trims all trailing whitespace characters in text files (spaces, tabs, no-break spaces, Mongolian vowel separators, en quads, em quads, en spaces, em spaces, three-per-em spaces, four-per-em spaces, six-per-em spaces, figure spaces, punctuation spaces, thin spaces, hair spaces, narrow no-break spaces, medium mathematical spaces, ideographic spaces, zero width spaces, and zero width no-break spaces).
 
             The <comment>--name, -a</comment> option transliterates and slugifies special characters in the directory name or file basename (filename part without extension). For example, foo bar.jpg -> foo-bar.jpg.
@@ -92,6 +97,7 @@ class CheckCommand extends Command
             new InputOption('leading-eol', 'l', InputOption::VALUE_NONE, 'Trim redundant leading newlines.'),
             new InputOption('middle-eol', 'm', InputOption::VALUE_OPTIONAL, 'Trim redundant middle empty newlines.', false),
             new InputOption('name', 'a', InputOption::VALUE_NONE, 'Fix file and directory names.'),
+            new InputOption('not', null, InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Skip given path(s)', []),
             new InputOption('permissions', 'u', InputOption::VALUE_NONE, 'Fix file and directory permissions.'),
             new InputOption('space-before-tab', 's', InputOption::VALUE_NONE, 'Clean spaces before tabs in the initial part of the line.'),
             new InputOption('trailing-whitespace', 'w', InputOption::VALUE_NONE, 'Trim trailing whitespace characters.'),
@@ -117,14 +123,30 @@ class CheckCommand extends Command
 
         $this->configurator->set($input->getOptions());
 
-        /** @var array<int,string> */
-        $paths = $input->getArgument('paths');
-
         $outputStyle = new OutputFormatterStyle('white', 'blue');
         $output->getFormatter()->setStyle('header', $outputStyle);
 
         /** @var FormatterHelper */
         $formatter = $this->getHelper('formatter');
+
+        /** @var array<int,string> */
+        $paths = [];
+        foreach ($input->getArgument('paths') as $path) {
+            if (file_exists($path)) {
+                $pathInfo = new File($path);
+                $paths[] = $pathInfo->getRealPath();
+            } else {
+                $formattedBlock = $formatter->formatBlock(
+                    ['Error:', sprintf('%s: No such file or directory', $path)],
+                    'error',
+                    true,
+                );
+
+                $output->writeln(['', $formattedBlock]);
+
+                return Command::FAILURE;
+            }
+        }
 
         $formattedBlock = $formatter->formatBlock(
             ['CHECKING', ...$paths],
@@ -134,9 +156,30 @@ class CheckCommand extends Command
 
         $output->writeln([$formattedBlock, '']);
 
+        // Set excluded paths.
+        $excludes = [];
+        foreach ($paths as $path) {
+            foreach ($input->getOption('not') as $exclude) {
+                if (file_exists($path . '/' . $exclude)) {
+                    $excludeInfo = new File($path . '/' . $exclude);
+                } elseif (file_exists($exclude)) {
+                    $excludeInfo = new File($exclude);
+                } else {
+                    continue;
+                }
+
+                $excludes[] = $excludeInfo->getRealPath();
+            }
+        }
+        $excludes = array_unique($excludes);
+
         $count = 0;
         foreach ($paths as $path) {
-            $iterator = $this->normalize($path, $output);
+            $iterator = $this->normalize(
+                $path,
+                $excludes,
+                $output,
+            );
             $count += iterator_count($iterator);
         }
 
@@ -147,11 +190,11 @@ class CheckCommand extends Command
             round(memory_get_peak_usage() / 1024 / 1024, 3),
         )]);
 
-        if (0 < count($this->logger->getAllLogs()) + count($this->logger->getAllErrors())) {
+        if (0 < $this->logger->getCounter()) {
             $formattedBlock = $formatter->formatBlock(
                 [sprintf(
                     '%d of %d %s should be fixed.',
-                    count($this->logger->getAllLogs()) + count($this->logger->getAllErrors()),
+                    $this->logger->getCounter(),
                     $count,
                     (1 === $count) ? 'file' : 'files',
                 )],
@@ -181,10 +224,7 @@ class CheckCommand extends Command
             $exitCode = Command::FAILURE;
         }
 
-        if (
-            0 < count($this->logger->getAllErrors())
-            || 0 < count($this->logger->getAllLogs())
-        ) {
+        if (0 < $this->logger->getCounter()) {
             $exitCode = Command::FAILURE;
         }
 
@@ -197,9 +237,20 @@ class CheckCommand extends Command
     /**
      * Run normalizator on given path and send output.
      */
-    private function normalize(string $path, OutputInterface $output): Iterator
-    {
-        $iterator = $this->finder->getTree($path);
+    private function normalize(
+        string $path,
+        array $excludes,
+        OutputInterface $output,
+    ): Iterator {
+        $iterator = $this->finder->getTree($path, static function (File $file) use ($excludes) {
+            foreach ($excludes as $exclude) {
+                if ($file->getRealPath() === $exclude) {
+                    return false;
+                }
+            }
+
+            return true;
+        });
 
         foreach ($iterator as $file) {
             $this->normalizator->normalize($file);